refacto: split adapter
