+++ external-import/tenable-security-center/tenable_security_center/adapters/tenable_security_center/v5_13_from_asset.py
@@ -4,18 +4,16 @@
 # isort is removing the type ignore untyped import comment conflicting with mypy
 
 import datetime
-import re
 from concurrent.futures import ThreadPoolExecutor
 from functools import lru_cache, partial
 from threading import Lock
-from typing import TYPE_CHECKING, Any, Callable, Iterable, Optional
+from typing import TYPE_CHECKING, Any, Iterable, Optional
 from urllib.parse import urlencode
 
 from pycti import (  # type: ignore[import-untyped]
     __version__ as pycti_version,  # pycti does not provide stubs
 )
 
-from pydantic import BaseModel, ConfigDict, Field, AwareDatetime, ValidationError
 from requests import HTTPError
 from semver import VersionInfo
 from tenable.errors import (  # type: ignore[import-untyped]
@@ -39,343 +37,21 @@ from tenable_security_center.ports.errors import (
     FindingRetrievalError,
 )
 
+from tenable_security_center.adapters.tenable_security_center.v5_13_common import (
+    FindingPydanticModel,
+    AssetPydanticModel,
+    FlattenRawResponseInSnakeCase,
+    _CVEAPI,
+    ASSETS_CHUNK_SIZE,
+    ASSET_FIELDS,
+    FINDINGS_CHUNK_SIZE,
+    FINDING_SEVERITIES_MAP,
+)
+
 
 if TYPE_CHECKING:
     from requests import Response
     from tenable_security_center.utils import AppLogger
-
-
-ASSETS_CHUNK_SIZE = 10
-FINDINGS_CHUNK_SIZE = 100
-
-FINDING_FILTERS = (
-    "acceptedRisk",
-    "acceptRiskStatus",
-    "assetID",
-    "auditFileID",
-    "benchmarkName",
-    "cceID",
-    "cpe",
-    "cveID",
-    "baseCVSSScore",
-    "cvssVector",
-    "cvssV3BaseScore",
-    "cvssV3Vector",
-    "dataFormat",
-    "daysMitigated",
-    "daysToMitigated",
-    "dnsName",
-    "exploitAvailable",
-    "exploitFrameworks",
-    "familyID",
-    "firstSeen",
-    "iavmID",
-    "ip",
-    "lastMitigated",
-    "lastSeen",
-    "mitigatedStatus",
-    "msbulletinID",
-    "operatingSystem",
-    "outputAssets",
-    "patchPublished",
-    "pluginModified",
-    "pluginPublished",
-    "pluginID",
-    "pluginName",
-    "pluginText",
-    "pluginType",
-    "policyID",
-    "port",
-    "protocol",
-    "recastRisk",
-    "recastRiskStatus",
-    "repositoryIDs",
-    "responsibleUserIDs",
-    "vulnRoutedUserIDs",
-    "vulnRoutingRuleID",
-    "severity",
-    "solutionID",
-    "stigSeverity",
-    "tcpport",
-    "udpport",
-    "uuid",
-    "vulnPublished",
-    "seolDate",
-    "wasMitigated",
-    "xref",
-    "vprScore",
-    "assetCriticalityRating",
-    "hostUUID",
-    "vulnUUID",
-    "assetExposureScore",
-    "aesSeverity",
-    "netbiosName",
-    "wasInputName",
-    "wasInputType",
-    "wasURL",
-    "wasHttpMethod",
-    "wasVuln",
-)  # Undocumented at that time in Tenable doc. This has been extractred from API response error message.
-
-FINDING_SEVERITIES_MAP = {
-    "info": 0,
-    "low": 1,
-    "medium": 2,
-    "high": 3,
-    "critical": 4,
-}
-
-ASSET_FIELDS = [
-    "id",
-    "uuid",
-    "tenableUUID",
-    "name",
-    "ipAddress",
-    "os",
-    "firstSeen",
-    "lastSeen",
-    "macAddress",
-    "source",
-    "repID",
-    "netBios",
-    "netBiosWorkgroup",
-    "createdTime",
-    "modifiedTime",
-]
-
-
-class _BaseModelWithoutExtra(BaseModel):
-    """Base model which does not accept other fields than the ones defined."""
-
-    model_config = ConfigDict(
-        extra="forbid",
-    )
-
-
-_camel_case_pattern = re.compile(r"([a-z])([A-Z])")
-
-
-class _FlattenRawResponseInSnakeCase:
-    """Flatten a raw response from the Tenable Security Center API and convert camel case key to snake case.
-
-    Examples:
-        >>> raw_response = {
-        ...     "fooBar": "Foo bar",
-        ...     "test": {"toTo": 1, "tiTi": 2}
-        ...     }
-        >>> flatten = _FlattenRawResponseInSnakeCase(raw_response)
-        >>> flatten.values["foo_bar"]
-        "Foo bar"
-        >>> flatten.values["test.to_to"]
-        1
-
-    """
-
-    def __init__(self, raw_response: dict[str, Any], sep: str = "."):
-        self.__raw_response = raw_response.copy()
-        self.__sep = sep
-        self.values = self.__process()
-
-    @staticmethod
-    def __cammel_to_snake_case(v: str) -> str:
-        return _camel_case_pattern.sub(r"\1_\2", v).lower()
-
-    @staticmethod
-    def __flatten_dict(
-        d: dict[str, Any],
-        sep: str = ".",
-        _parent_key: str = "",
-    ) -> dict[str, Any]:
-        """Flattens a dictionary of nested dictionnaries and convert camel case key to snake case.
-
-        Args:
-            d(dict[str, Any]): The dictionary to flatten.
-            sep(str): The separator to use between keys.
-            _parent_key(str): The parent key to use (for recursion only, it should not be passed).
-
-        Returns:
-            dict[str, Any]: The flatten dictionary.
-
-        """
-        items: list[tuple[str, Any]] = []
-        _parent_key_snake = (
-            _FlattenRawResponseInSnakeCase.__cammel_to_snake_case(str(_parent_key))
-            if _parent_key
-            else ""
-        )
-        for k, v in d.items():
-            k_snake = _FlattenRawResponseInSnakeCase.__cammel_to_snake_case(str(k))
-            new_key = f"{_parent_key_snake}{sep}{k_snake}" if _parent_key else k_snake
-            if isinstance(v, dict):  # If the value is a dictionary, recurse
-                items.extend(
-                    _FlattenRawResponseInSnakeCase.__flatten_dict(
-                        d=v, _parent_key=new_key, sep=sep
-                    ).items()
-                )
-            else:
-                items.append((new_key, v))
-        return dict(items)
-
-    def __process(self) -> dict[str, Any]:
-        return self.__flatten_dict(self.__raw_response, sep=self.__sep)
-
-    def apply(
-        self,
-        func: Callable[..., Any],
-        key: str,
-        optional: bool = False,
-        other_excepted_values: Optional[list[str]] = None,
-    ) -> Any:
-        """Apply a function to a given field."""
-        val = self.values.get(key)
-        _other_excepted_values = other_excepted_values or []
-        if optional and (not (val) or val in _other_excepted_values):
-            return None
-        return func(val)
-
-    def int_field(self, key: str, optional: bool = False) -> Optional[int]:
-        """Apply a int casting to a given field.
-
-        Args:
-            key(str): The key to select the field.
-            optional(bool): If the field is optional.
-
-        Returns:
-            Optional[int]: The value as an integer.
-
-        """
-        value = self.apply(int, key, optional=optional)
-        return value if value is not None else None
-
-    def bool_int_field(self, key: str, optional: bool = False) -> Optional[bool]:
-        """Apply a bool casting to a given field."""
-        value = self.apply(lambda x: bool(int(x)), key, optional=optional)
-        return value if value is not None else None
-
-    def float_field(self, key: str, optional: bool = False) -> Optional[float]:
-        """Apply a float casting to a given field."""
-        value = self.apply(float, key, optional=optional)
-        return value if value is not None else None
-
-    def timestamp_field(self, key: str, optional: bool) -> Optional[datetime.datetime]:
-        """Apply a timestamp casting to a given field."""
-        dt: Optional[datetime.datetime] = self.apply(
-            lambda x: datetime.datetime.fromtimestamp(int(x)),
-            key,
-            optional=optional,
-            other_excepted_values=["-1"],
-        )
-        if dt is not None:
-            # make utc time aware if no timezone
-            if dt.tzinfo is None:
-                dt = dt.replace(tzinfo=datetime.timezone.utc)
-        return dt
-
-    def str_list_field(
-        self, key: str, optional: bool, sep: str = ","
-    ) -> Optional[list[Any]]:
-        """Apply a split to a given field."""
-        str_list = self.apply(lambda x: x.split(sep), key, optional=optional)
-        return list(str_list) if str_list else None
-
-
-class CVEPydanticModel(_BaseModelWithoutExtra):
-    """Pydantic model for a CVE from the Tenable Security Center API."""
-
-    name: str = Field(...)
-    description: str = Field(...)
-    publication_datetime: AwareDatetime = Field(...)
-    last_modified_datetime: AwareDatetime = Field(...)
-    cpes: Optional[list[str]] = Field(None)
-    cvss_v3_score: Optional[float] = Field(None)
-    cvss_v3_vector: Optional[str] = Field(None)
-    epss_score: Optional[float] = Field(None)
-    epss_percentile: Optional[float] = Field(None)
-
-
-class _CVEAPI(CVEPort):
-    """Represent a CVE from the Tenable Security Center API."""
-
-    def __init__(self, **data: dict[str, Any]):
-        try:
-            self._pydantic_model = CVEPydanticModel.model_validate(data)
-        except ValidationError as e:
-            raise CVERetrievalError(
-                f"Error while validating the CVE data: {data}"
-            ) from e
-
-    @property
-    def name(self) -> str:
-        return self._pydantic_model.name
-
-    @property
-    def description(self) -> str:
-        return self._pydantic_model.description
-
-    @property
-    def publication_datetime(self) -> datetime.datetime:
-        return self._pydantic_model.publication_datetime
-
-    @property
-    def last_modified_datetime(self) -> datetime.datetime:
-        return self._pydantic_model.last_modified_datetime
-
-    @property
-    def cpes(self) -> Optional[list[str]]:
-        return self._pydantic_model.cpes
-
-    @property
-    def cvss_v3_score(self) -> Optional[float]:
-        return self._pydantic_model.cvss_v3_score
-
-    @property
-    def cvss_v3_vector(self) -> Optional[str]:
-        return self._pydantic_model.cvss_v3_vector
-
-    @property
-    def epss_score(self) -> Optional[float]:
-        return self._pydantic_model.epss_score
-
-    @property
-    def epss_percentile(self) -> Optional[float]:
-        return self._pydantic_model.epss_percentile
-
-    @classmethod
-    def from_raw_response(cls, raw_response: dict[str, Any]) -> "_CVEAPI":
-        """Create a CVE from a raw response."""
-        # Note: The values might be None, specific validations will be performed by pydantic.
-        name = raw_response.get("primary_vuln_id")
-        description = raw_response.get("descriptions", [{}])[0].get("description_text")
-        publication_datetime = raw_response.get("descriptions", [{}])[0].get(
-            "publication_date"
-        )
-        last_modified_datetime = raw_response.get("descriptions", [{}])[-1].get(
-            "publication_date"
-        )
-        cpes = [
-            str(uri)  # explicit cast for mypy linter
-            for uri in raw_response.get("cpe_metrics", [])
-            if str(uri).startswith("cpe:") or str(uri).startswith("p-cpe:")
-        ]
-        cvss_v3_score = raw_response.get("cvss_metrics", [{}])[0].get(
-            "cvss3_base_score"
-        )
-        cvss_v3_vector = raw_response.get("cvss_metrics", [{}])[0].get(
-            "cvss3_base_vector"
-        )
-        epss_score = raw_response.get("epss_metrics", [{}])[0].get("epss")
-        epss_percentile = raw_response.get("epss_metrics", [{}])[0].get("percentile")
-        return cls(
-            name=name,  # type: ignore[arg-type]
-            description=description,
-            publication_datetime=publication_datetime,
-            last_modified_datetime=last_modified_datetime,
-            cpes=cpes,  # type: ignore[arg-type]
-            cvss_v3_score=cvss_v3_score,
-            cvss_v3_vector=cvss_v3_vector,
-            epss_score=epss_score,
-            epss_percentile=epss_percentile,
-        )
 
 
 class _CVEsAPI:  # pylint: disable=too-few-public-methods
@@ -422,69 +98,6 @@ class _CVEsAPI:  # pylint: disable=too-few-public-methods
         """Fetch and process the CVEs."""
         for raw_cve in self._fetch_data_chunk(cve_ids):
             yield _CVEAPI.from_raw_response(raw_cve)
-
-
-class FindingPydanticModel(_BaseModelWithoutExtra):
-    """Pydantic model for a finding from the Tenable Security Center API.
-
-    Notes :
-        Used to validate API results with understandable Validation exception.
-    """
-
-    plugin_name: str = Field(...)
-    cve: Optional[list[str]] = Field(None)
-    cpe: Optional[list[str]] = Field(None)
-    plugin_id: str = Field(...)
-    has_been_mitigated: bool = Field(...)
-    accept_risk: bool = Field(...)
-    recast_risk: bool = Field(...)
-    ip: str = Field(...)
-    uuid: Optional[str] = Field(None)
-    port: int = Field(...)
-    protocol: str = Field(...)
-    first_seen: AwareDatetime = Field(...)
-    last_seen: AwareDatetime = Field(...)
-    exploit_available: bool = Field(...)
-    exploit_ease: Optional[str] = Field(None)
-    exploit_frameworks: Optional[list[str]] = Field(None)
-    synopsis: Optional[str] = Field(None)
-    description: Optional[str] = Field(None)
-    solution: Optional[str] = Field(None)
-    see_also: Optional[list[str]] = Field(None)
-    risk_factor: Optional[str] = Field(None)
-    stig_severity: Optional[str] = Field(None)
-    severity_name: str = Field(...)
-    vpr_score: Optional[float] = Field(None)
-    vpr_context: Optional[list[str]] = Field(None)
-    base_score: Optional[float] = Field(None)
-    temporal_score: Optional[float] = Field(None)
-    cvss_vector: Optional[str] = Field(None)
-    cvss_v3_base_score: Optional[float] = Field(None)
-    cvss_v3_temporal_score: Optional[float] = Field(None)
-    cvss_v3_vector: Optional[str] = Field(None)
-    vuln_pub_date: Optional[AwareDatetime] = Field(None)
-    patch_pub_date: Optional[AwareDatetime] = Field(None)
-    plugin_pub_date: Optional[AwareDatetime] = Field(None)
-    plugin_mod_date: Optional[AwareDatetime] = Field(None)
-    check_type: Optional[str] = Field(None)
-    version: Optional[str] = Field(None)
-    bid: Optional[list[str]] = Field(None)
-    xref: Optional[list[str]] = Field(None)
-    seol_date: AwareDatetime = Field(...)
-    plugin_text: Optional[str] = Field(None)
-    dns_name: Optional[str] = Field(None)
-    mac_address: Optional[str] = Field(None)
-    netbios_name: Optional[str] = Field(None)
-    operating_system: Optional[str] = Field(None)
-    recast_risk_rule_comment: Optional[list[str]] = Field(None)
-    accept_risk_rule_comment: Optional[list[str]] = Field(None)
-    host_uniqueness: list[str] = Field(...)
-    host_uuid: Optional[str] = Field(None)
-    acr_score: Optional[float] = Field(None)
-    asset_exposure_score: float = Field(...)
-    vuln_uniqueness: list[str] = Field(...)
-    vuln_uuid: Optional[str] = Field(None)
-    uniqueness: list[str] = Field(...)
 
 
 class _FindingAPI(FindingPort):
@@ -727,7 +340,7 @@ class _FindingAPI(FindingPort):
     @staticmethod
     def parse_response(raw_response: dict[str, Any]) -> dict[str, Any]:
         """Parse finding needed arguments from a raw response."""
-        flatten_response_snake_case_key = _FlattenRawResponseInSnakeCase(
+        flatten_response_snake_case_key = FlattenRawResponseInSnakeCase(
             raw_response, sep="_"
         )
 
@@ -989,23 +602,6 @@ class _FindingsAPI:
             )
 
 
-class AssetPydanticModel(_BaseModelWithoutExtra):
-    """Pydantic model for an asset from the Tenable Security Center API."""
-
-    id: str = Field(...)
-    uuid: str = Field(...)
-    tenable_uuid: Optional[str] = Field(None)
-    name: str = Field(...)
-    operating_systems: Optional[list[str]] = Field(None)
-    first_seen: datetime.datetime = Field(...)
-    last_seen: datetime.datetime = Field(...)
-    mac_address: Optional[str] = Field(None)
-    created_time: datetime.datetime = Field(...)
-    modified_time: datetime.datetime = Field(...)
-    ip_address: str = Field(...)
-    repository_id: str = Field(...)
-
-
 class _AssetAPI(AssetPort):
     """Represents an asset from the Tenable Security Center API."""
 
@@ -1087,7 +683,7 @@ class _AssetAPI(AssetPort):
         optional_list_fields = [("operating_systems", ",")]
 
         datetime_fields = ["first_seen", "last_seen", "created_time", "modified_time"]
-        flatten_response_snake_case_key = _FlattenRawResponseInSnakeCase(
+        flatten_response_snake_case_key = FlattenRawResponseInSnakeCase(
             raw_response, sep="_"
         )
 
@@ -1119,6 +715,58 @@ class _AssetAPI(AssetPort):
             }
         )
         return data
+
+
+class _ScanResultsAPI:
+    def __init__(
+        self,
+        tsc_client: TenableSC,
+        logger: "AppLogger",
+        since_datetime: datetime.datetime,
+    ):
+        self.logger = logger
+        self.client = tsc_client
+        self.since_datetime = since_datetime
+
+    def get_completed_scan_ids(self) -> list[str]:
+        params = {
+            "filters": "usable,completed,optimizeCompletedScans",
+            "fields": "id",
+            "timeComparedField": "createdTime",
+            "startTime": int(self.since_datetime.timestamp()),
+        }
+        url = f"scanResult?{'&'.join([f'{k}={v}' for k, v in params.items()])}"
+        try:
+            resp_scan_results = self.client.get(url)
+            resp_scan_results.raise_for_status()
+            json_resp_scan_results = resp_scan_results.json()
+            return [scan["id"] for scan in json_resp_scan_results["response"]["usable"]]
+        except HTTPError as e:
+            self.logger.error(
+                f"Error while fetching data from Tenable Security Center: {e}"
+            )
+            raise AssetRetrievalError(
+                "Error while fetching data from Tenable Security Center."
+            ) from e
+
+    def get_scanned_assets_info(self, scan_id: str) -> tuple[str, str]:
+        fields = ["id", "name", "repository", "progress"]
+        url = f"scanResult/{scan_id}?fields={','.join(fields)}"
+        try:
+            resp_scan_result = self.client.get(url)
+            resp_scan_result.raise_for_status()
+            json_data = resp_scan_result.json()["response"]
+            if not json_data["progress"]:  # could be an empty list, empty dict or null
+                return "", ""
+            return json_data["progress"]["scannedIPs"], json_data["repository"]["id"]
+
+        except (HTTPError, IndexError) as e:
+            self.logger.error(
+                f"Error while fetching data from Tenable Security Center: {e}"
+            )
+            raise AssetRetrievalError(
+                "Error while fetching data from Tenable Security Center."
+            ) from e
 
 
 class _AssetsChunkAPI(AssetsChunkPort):
@@ -1261,17 +909,33 @@ class AssetsAPI(AssetsPort):
             num_threads=self.num_threads,
             cves_api=self._cves_api,
         )
+        self._scan_results_api: _ScanResultsAPI = _ScanResultsAPI(
+            tsc_client=self.client,
+            logger=self.logger,
+            since_datetime=self.since_datetime,
+        )
 
     def _fetch_data_chunks(self) -> Iterable[_AssetsChunkAPI]:
         """Fetch all data chunks from the API."""
+        # As we can't filter assets by last seen date, we use scan results endpoints to filter
+        # on assets that have been scanned since the provided datetime, then we fetch the assets using
+        # these information.
+        self.logger.debug("Fetching scan results from Tenable Security Center.")
+        scan_ids = self._scan_results_api.get_completed_scan_ids()
+        assets_info = [
+            self._scan_results_api.get_scanned_assets_info(scan_id)
+            for scan_id in scan_ids
+        ]
         filters = [
+            ("ip", "=", ",".join({ip for ip, _ in assets_info if ip})),
             (
-                "lastSeen",
+                "repositoryAll",
                 "=",
-                f"{int(self.since_datetime.timestamp())}-{int(datetime.datetime.now().timestamp())}",
-            )  # ts_start-ts_end
-        ]  # this lastSeen Filter does not do anything. We will reject assets with lastSeen < Since_date manually.
-
+                ",".join(
+                    {repository_id for _, repository_id in assets_info if repository_id}
+                ),
+            ),
+        ]
         # Only get "id" field to get total_records
         total_records = int(
             _AssetsChunkAPI(  # pylint: disable=protected-access
@@ -1288,7 +952,7 @@ class AssetsAPI(AssetsPort):
             .get("totalRecords", 0)
         )
 
-        self.logger.debug(
+        self.logger.info(
             f"Fetching {total_records} assets from Tenable Security Center."
         )
 
@@ -1314,84 +978,3 @@ class AssetsAPI(AssetsPort):
     def chunks(self) -> Iterable[_AssetsChunkAPI]:
         """Fetch and yield the assets chunks."""
         yield from self._fetch_data_chunks()
-
-
-# if __name__ == "__main__":
-#     from tenable.sc import TenableSC
-
-#     f = _FindingAPI.from_raw_response(
-#         {
-#     "pluginID": "10107",
-#     "severity": {
-#       "id": "0",
-#       "name": "Info",
-#       "description": "Informative"
-#     },
-#     "hasBeenMitigated": "1",
-#     "acceptRisk": "0",
-#     "recastRisk": "0",
-#     "ip": "10.238.64.1",
-#     "uuid": "",
-#     "port": "443",
-#     "protocol": "TCP",
-#     "pluginName": "HTTP Server Type and Version",
-#     "firstSeen": "1570291573",
-#     "lastSeen": "1729354189",
-#     "exploitAvailable": "No",
-#     "exploitEase": "",
-#     "exploitFrameworks": "",
-#     "synopsis": "A web server is running on the remote host.",
-#     "description": "This plugin attempts to determine the type and the version of the   remote web server.",
-#     "solution": "",
-#     "seeAlso": "",
-#     "riskFactor": "None",
-#     "stigSeverity": "",
-#     "vprScore": "",
-#     "vprContext": "[]",
-#     "baseScore": "",
-#     "temporalScore": "",
-#     "cvssVector": "",
-#     "cvssV3BaseScore": "",
-#     "cvssV3TemporalScore": "",
-#     "cvssV3Vector": "",
-#     "cpe": "",
-#     "vulnPubDate": "-1",
-#     "patchPubDate": "-1",
-#     "pluginPubDate": "946987200",
-#     "pluginModDate": "1604059200",
-#     "checkType": "remote",
-#     "version": "1.141",
-#     "cve": "",
-#     "bid": "",
-#     "xref": "IAVT #0001-T-0931",
-#     "seolDate": "-1",
-#     "pluginText": "<plugin_output>The remote web server type is :\n\nopenresty</plugin_output>",
-#     "dnsName": "_gateway.incus",
-#     "macAddress": "00:16:3e:a1:12:f7",
-#     "netbiosName": "",
-#     "operatingSystem": "Linux Kernel 2.6",
-#     "ips": "10.238.64.1",
-#     "recastRiskRuleComment": "",
-#     "acceptRiskRuleComment": "",
-#     "hostUniqueness": "repositoryID,ip,dnsName",
-#     "hostUUID": "",
-#     "acrScore": "6.0",
-#     "keyDrivers": "{\"internet exposure\":\"internal\",\"device capability\":\"dns_server\",\"device type\":\"general_purpose\"}",
-#     "assetExposureScore": "547",
-#     "vulnUniqueness": "repositoryID,ip,port,protocol,pluginID",
-#     "vulnUUID": "",
-#     "uniqueness": "repositoryID,ip,dnsName",
-#     "family": {
-#       "id": "11",
-#       "name": "Web Servers",
-#       "type": "active"
-#     },
-#     "repository": {
-#       "id": "1",
-#       "name": "Live",
-#       "description": "",
-#       "dataFormat": "IPv4"
-#     },
-#     "pluginInfo": "10107 (443/6) HTTP Server Type and Version"
-#   }
-#     )
